Split out display to its own file
